+++ src/app/modules/cart/services/cart.service.ts
@@ -1,18 +1,34 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Cart } from '../model/cart';
-import { map, Observable, of, switchMap, tap } from 'rxjs';
+import {
+  catchError,
+  map,
+  Observable,
+  of,
+  switchMap,
+  tap,
+  throwError,
+} from 'rxjs';
 import { UserService } from '../../user/services/user.service';
-import { parse } from 'path';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from '../../../shared/confirm-dialog/confirm-dialog.component';
+import { Product } from '../../products/models/product';
 
 @Injectable({
   providedIn: 'root',
 })
 export class CartService {
   private serverUrl = 'http://localhost:3000/carts';
+  private productServerUrl = 'http://localhost:3000/products';
   status: string = 'added';
+  forCheckOut: boolean = true;
 
-  constructor(private http: HttpClient, private userService: UserService) {}
+  constructor(
+    private http: HttpClient,
+    private userService: UserService,
+    private dialog: MatDialog
+  ) {}
 
   getCurrentUserId = () => {
     const user = this.userService.getCurrentUser();
@@ -74,7 +90,7 @@ export class CartService {
     return this.getCarts().pipe(
       map((carts) => {
         if (carts.length === 0) return 0;
-        return carts.reduce((total, cart) => total + (cart.price), 0);
+        return carts.reduce((total, cart) => total + cart.price, 0);
       })
     );
   }
@@ -132,15 +148,33 @@ export class CartService {
 
   incrementCount(id: string): Observable<Cart> {
     const url = `${this.serverUrl}/${id}`;
+
     return this.http.get<Cart>(url).pipe(
-      map((cart) => {
-        cart.price = cart.price / cart.quantity;
-        cart.quantity += 1;
-        cart.price *= cart.quantity;
-        return cart;
+      switchMap((cart) => {
+        const productUrl = `${this.productServerUrl}/${cart.productId}`;
+
+        return this.http.get<Product>(productUrl).pipe(
+          map((product) => {
+            if (cart.quantity + 1 > product.quantity) {
+              throw new Error('Exceeds available product stock');
+            }
+
+            const updatedCart = { ...cart };
+            updatedCart.quantity += 1;
+            updatedCart.price =
+              (cart.price / cart.quantity) * updatedCart.quantity;
+
+            return updatedCart;
+          }),
+          switchMap((updatedCart) => this.http.patch<Cart>(url, updatedCart)),
+          tap(() => console.log(`Added quantity for cart ${id}`))
+        );
       }),
-      switchMap((updateCart) => this.http.patch<Cart>(url, updateCart)),
-      tap((updatedCart) => console.log(`Added quantity for cart ${id}`))
+      catchError((error) => {
+        console.error('Error updating cart quantity:', error);
+        alert('Exceeds product stock.');
+        return throwError(() => error);
+      })
     );
   }
 
@@ -152,9 +186,23 @@ export class CartService {
         cart.quantity -= 1;
 
         if (cart.quantity === 0) {
-          return this.deleteItemFromCart(id).pipe(
-            tap(() => console.log('Quantity is zero')),
-            map(() => null)
+          const message = `Are you sure you want to delete this item from the cart?`;
+
+          const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+            data: { message },
+          });
+
+          return dialogRef.afterClosed().pipe(
+            switchMap((result) => {
+              if (result) {
+                return this.deleteItemFromCart(id).pipe(
+                  tap(() => console.log('Item deleted')),
+                  map(() => null)
+                );
+              } else {
+                return of(cart);
+              }
+            })
           );
         }
         cart.price *= cart.quantity;
@@ -172,25 +220,41 @@ export class CartService {
 
   editQuantity(id: string, quantity: number): Observable<Cart> {
     const url = `${this.serverUrl}/${id}`;
-    return this.http.get<Cart>(url).pipe(
-      map((cart) => {
-        cart.price = cart.price / cart.quantity;
-        cart.quantity = quantity;
+    quantity = Math.floor(quantity);
 
-        if (cart.quantity === 0) {
-          return this.deleteItemFromCart(id).pipe(
-            tap(() => console.log('Quantity is zero')),
-            map(() => null)
-          );
-        }
-        cart.price *= cart.quantity;
-        return this.http
-          .patch<Cart>(url, cart)
-          .pipe(
-            tap((updatedCart) => console.log(`Edit quantity for cart ${id}`))
-          );
+    return this.http.get<Cart>(url).pipe(
+      switchMap((cart) => {
+        const productUrl = `${this.productServerUrl}/${cart.productId}`;
+
+        return this.http.get<Product>(productUrl).pipe(
+          map((product) => {
+            if (quantity > product.quantity) {
+              this.forCheckOut = false;
+              throw new Error('Exceeds available product stock');
+            }
+
+            const updatedCart = { ...cart };
+            updatedCart.quantity = quantity;
+
+            updatedCart.price =
+              (cart.price / cart.quantity) * updatedCart.quantity;
+
+            this.forCheckOut = true;
+
+            return updatedCart;
+          }),
+          switchMap((updatedCart) =>
+            this.http
+              .patch<Cart>(url, updatedCart)
+              .pipe(tap(() => console.log(`Edited quantity for cart ${id}`)))
+          )
+        );
       }),
-      switchMap((result) => result as Observable<Cart>)
+      catchError((error) => {
+        console.error('Error updating cart quantity:', error);
+        alert('Exceeds product stock.');
+        return throwError(() => error);
+      })
     );
   }
 
@@ -201,8 +265,7 @@ export class CartService {
   }
 
   getPendingCarts = () => {
-
-    let statusPending = "pending"
+    let statusPending = 'pending';
     const currentUserId = this.getCurrentUserId();
 
     if (currentUserId === null || currentUserId === '0') {
@@ -221,14 +284,13 @@ export class CartService {
         console.log('Fetched and filtered carts:', filteredCarts)
       )
     );
-
-  }
+  };
 
   pendingStatus(id: string): Observable<Cart> {
     const url = `${this.serverUrl}/${id}`;
     return this.http.get<Cart>(url).pipe(
       map((cart) => {
-        cart.status = "pending"
+        cart.status = 'pending';
         return cart;
       }),
       switchMap((updateCart) => this.http.patch<Cart>(url, updateCart)),
